model test improvements/parameterizations
